requirements.txt
flask
gql
requests
python-dotenv